Couldn't get project to load.
Was using a new version of monogame and couldn't get it working. Removed
the newer version and installed 3.0.(it's what the book uses). After
installing opal, it works as expected! Was hoping to get some of my own
code down but it's at least ready to rock. That's it for tonight.