InputWrapper works so far.
Need to clean it up and move mouse implimentation into the wrapper
class. Ended up copying over the previous projects game1.cs into this
project and make a few tweaks to it to get it to work as expected. Seems
to work for the most part. Poles for thumbsticks and up/down are
reversed.....Might should fix that.