Mouse support added to InputWrapper
Got it working for the most part. Need to clean up a few items and
figure out the best way to deal with calling mouse.state() so many
times. Got to be a better way to get the vector data as well. Maybe call
it with out keyword? It's a bit unusual for an input device as it
provides mouse position as well as button data.